requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langgraph>=0.0.20
supabase>=2.3.0

# Document processing
pypdf>=3.17.0
docx2txt>=0.8
python-pptx>=0.6.21
beautifulsoup4>=4.12.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web framework (for demo)
streamlit>=1.28.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0